Module aws_sdk_migrationhub::error

Modules

access_denied_exception

See AccessDeniedException

dry_run_operation

See DryRunOperation

home_region_not_set_exception

See HomeRegionNotSetException

internal_server_error

See InternalServerError

invalid_input_exception

See InvalidInputException

policy_error_exception

See PolicyErrorException

resource_not_found_exception

See ResourceNotFoundException

service_unavailable_exception

See ServiceUnavailableException

throttling_exception

See ThrottlingException

unauthorized_operation

See UnauthorizedOperation

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

AssociateCreatedArtifactError

AssociateDiscoveredResourceError

CreateProgressUpdateStreamError

DeleteProgressUpdateStreamError

DescribeApplicationStateError

DescribeMigrationTaskError

DisassociateCreatedArtifactError

DisassociateDiscoveredResourceError

DryRunOperation

Exception raised to indicate a successfully authorized action when the

HomeRegionNotSetException

The home region is not set. Set the home region to continue.

ImportMigrationTaskError

InternalServerError

Exception raised when an internal, configuration, or dependency error is

InvalidInputException

Exception raised when the provided input violates a policy constraint or is entered in

ListApplicationStatesError

ListCreatedArtifactsError

ListDiscoveredResourcesError

ListMigrationTasksError

ListProgressUpdateStreamsError

NotifyApplicationStateError

NotifyMigrationTaskStateError

PolicyErrorException

Exception raised when there are problems accessing Application Discovery Service

PutResourceAttributesError

ResourceNotFoundException

Exception raised when the request references a resource (Application Discovery Service

ServiceUnavailableException

Exception raised when there is an internal, configuration, or dependency error

ThrottlingException

The request was denied due to request throttling.

UnauthorizedOperation

Exception raised to indicate a request was not authorized when the DryRun

Enums

AssociateCreatedArtifactErrorKind

AssociateDiscoveredResourceErrorKind

CreateProgressUpdateStreamErrorKind

DeleteProgressUpdateStreamErrorKind

DescribeApplicationStateErrorKind

DescribeMigrationTaskErrorKind

DisassociateCreatedArtifactErrorKind

DisassociateDiscoveredResourceErrorKind

ImportMigrationTaskErrorKind

ListApplicationStatesErrorKind

ListCreatedArtifactsErrorKind

ListDiscoveredResourcesErrorKind

ListMigrationTasksErrorKind

ListProgressUpdateStreamsErrorKind

NotifyApplicationStateErrorKind

NotifyMigrationTaskStateErrorKind

PutResourceAttributesErrorKind