Module aws_sdk_migrationhub::model

Modules

application_state

See ApplicationState

created_artifact

See CreatedArtifact

discovered_resource

See DiscoveredResource

migration_task

See MigrationTask

migration_task_summary

See MigrationTaskSummary

progress_update_stream_summary

See ProgressUpdateStreamSummary

resource_attribute

See ResourceAttribute

task

See Task

Structs

ApplicationState

The state of an application discovered through Migration Hub import, the AWS Agentless

CreatedArtifact

An ARN of the AWS cloud resource target receiving the migration (e.g., AMI, EC2

DiscoveredResource

Object representing the on-premises resource being migrated.

MigrationTask

Represents a migration task in a migration tool.

MigrationTaskSummary

MigrationTaskSummary includes MigrationTaskName,

ProgressUpdateStreamSummary

Summary of the AWS resource used for access control that is implicitly linked to your

ResourceAttribute

Attribute associated with a resource.

Task

Task object encapsulating task information.

Enums

ApplicationStatus

ResourceAttributeType

Status