Module aws_sdk_migrationhub::operation

Structs

AssociateCreatedArtifact

Associates a created artifact of an AWS cloud resource, the target receiving the

AssociateDiscoveredResource

Associates a discovered resource ID from Application Discovery Service with a migration

CreateProgressUpdateStream

Creates a progress update stream which is an AWS resource used for access control as

DeleteProgressUpdateStream

Deletes a progress update stream, including all of its tasks, which was previously

DescribeApplicationState

Gets the migration status of an application.

DescribeMigrationTask

Retrieves a list of all attributes associated with a specific migration task.

DisassociateCreatedArtifact

Disassociates a created artifact of an AWS resource with a migration task performed by a

DisassociateDiscoveredResource

Disassociate an Application Discovery Service discovered resource from a migration

ImportMigrationTask

Registers a new migration task which represents a server, database, etc., being migrated

ListApplicationStates

Lists all the migration statuses for your applications. If you use the optional

ListCreatedArtifacts

Lists the created artifacts attached to a given migration task in an update stream. This

ListDiscoveredResources

Lists discovered resources associated with the given MigrationTask.

ListMigrationTasks

Lists all, or filtered by resource name, migration tasks associated with the user

ListProgressUpdateStreams

Lists progress update streams associated with the user account making this call.

NotifyApplicationState

Sets the migration state of an application. For a given application identified by the

NotifyMigrationTaskState

Notifies Migration Hub of the current status, progress, or other detail regarding a

PutResourceAttributes

Provides identifying details of the resource being migrated so that it can be associated